requirements.txt
tqdm==4.31.1
matplotlib==3.0.3
nltk==3.4
pandas==0.24.1
seaborn==0.9.0
scipy==1.2.1
annoy==1.15.1
numpy==1.16.2
requests==2.21.0
torch==1.0.1.post2
ipython=6.2.1
feedly-client==0.19
bs4==0.0.1